JournalClient: change the prefix init argument dafault value to None
and give this None value the special meaning of 'use the default value',
i.e. 'swh.journal.objects' (specifically, the value stored as
  swh.journal.DEFAULT_PREFIX).

This is required to make it easier to handle cli arguments and config
values read from the config file. Let the None value be the default
(from the cli config/argument handling) as long as possible and decide
only in the JournalClient what "default" really means.