replay: do not crash if process_replay_objects_content() fails on one object
Catch exceptions and log human readable ids of failed objects.

We do not want the whole process to stop if the copy of one object fails
(be it when reading or writing).

This is far from enough, we will have to deal with a better solution
(retry the requests to ward off temporary failures, better logging of failed
object ids, tooling for fixing these missed objects afterward...)