Be more careful with content generation in test_write_replay
When using the object_dicts() strategy, we often end up not generating any
contents; This works because the `replayer.process` call is within a loop,
guarded with the length of the queue.

If the queue is empty, `replayer.process` is never called, but the test is also
useless. So we add an assertion to that effect.