Rename JournalClient.max_messages to JournalClient.stop_after_objects
After various refactorings, the meaning of `max_messages` got muddled and we're
at a point where it doesn't mean anything anymore. `stop_after_objects` is
clearer as to what behavior the parameter is actually trying to achieve. We also
rename the replay command-line argument to the same name.

These refactorings also had us end up with a loop inside JournalClient.process,
while some callers still had a loop around it which ever got called once (or 0
times when the surrounding code was buggy). This commit removes all these outer
loops as well, keeping only the `JournalClient.process` inner loop.

While we're here, we use the opportunity to clarify and expand the documentation
of the JournalClient.