replayer: factor out legacy objects fixers
This will allow other journal clients to use the fixers, instead of this
being solely a replayer feature.

One could also imagine integrating the fixer directly in the journal so
that the fixes happen transparently for all the users without having to
call the fixer explicitely, and this diff is a first step in this
direction.