serializers: Make kafka_to_key implem compatible with stable version
The current debian storage build on stable [1] fails on an unexposed parameter.
The current debian journal build passed on stable probably because it missed a
test on that particular method.

This commit removes the need to turn off such parameter (default to True). It's
not needed for the kafka objects because it prevents the use of dict with null
keys. This is not the case for our current kafka objects.

This also allows to be compatible with current version onward. In effect,
fixing the failed build.

[1] https://jenkins.softwareheritage.org/view/Debian%20packages/job/debian/job/packages/job/DSTO/job/gbp-buildpackage/177/console