pytest_plugin: modify the kafka_server fixture so topics are created beforehand
Since we now check that topics exist on the kafka broker when instantiating a
JournalClient, topics but be created when the mock kafka broker is started
for tests to keep working (e.g. in swh.storage).

This needs a small adaptation in the JournalClient code itself to ensure
"empty" messages used to create topics will be silently ignored.

The kafka_server() fixture is split in 2 fixtures:
- kafka_server_base() that only creates the mock kafka server,
- kafka_server that() uses this later and creates the topics resulting from
  the conjunction of the kafka_prefix() and the new object_types() fixture.

This new object_types() fixture is thus used to define the list
of topics (in conjunction with the kafka_prefix fixture) a (mock)
kafka_server() will create at startup time.