pytest_plugin: remove '.swh.journal.objects' from test_config fixture
This induces a tigh coupling with code from other packages that use this
fixture (e.g. swh.storage currently have to modify the kafka_prefix resulting
from the fixture so the consumer and the writer are in sync).

If someone wants a customized version of a kafka_prefix, it must be done by
overriding the kafka_prefix fixture, not by modifying its returned value.

Note that this commit will break swh.storage.tests.test_kafka_writer until
it is updated.