Add support for a priviledged "channel" of topics for non-anonymized objects
The idea is to publish on a "public" prefix anonymized objects and have a
priviledged prefix for non-anonymized ones. Every model object which
".anonymize()" method returns a(n anonymized) model object will be sent
to both channels (anonymized version on the regular channel, and
original version on the privileged channel).

Currently only Release and Revision objects are anonymizable.

For anonymizable objects:

- regular (non-anonymized) objects are sent to the topic

  "{prefix}_privileged.{object_type}"

- anonymized version of these objects are sent to

  "{prefix}.{object_type}"

On the client side, a new boolean "privileged" config parameter is used to
select privileged topics if vivible on the Kafka broker at subscrition time.