Improve tests coverage by adding missing model types in journal_data
Improve the coverage of the test set provided in journal_data, namely
add directories, snapshots and skipped contents.

This requires to adapt a bit the function assert_all_objects_consumed() from
pytest_plugin, because we want to be able to handle anonymizable objects
specifically. So for now, simply add an 'exclude' argument.

It also needs the kafka_to_value serializer function to ensure tuples are
used instead of lists.