journal_data: Make origin-visit optional fields to None
This sets the fields status, metadata and snapshot from the origin-visit
entries to None (as per the recent model change D3340)

This pulls the creation of origin-visit-status with almost the same
origin-visit entries as before. The origin-visit-status dates are slightly
larger than the origin-visit ones so we are able to reflect the reality of
operations (first creation of origin visits, then creation of
origin-visit-status associated to that origin-visit).

Related to T2310