Refactor pytest_plugins.assert_all_objects_consumed() to use model objects
instead of dicts.

This is needed because we are beginning to use converters in swh.model,
so the dict used as input is not guaranteed to be the exact same one
as the one retrieved from the journal.

For this, we need to include the data attribute in journal_data's
Content objects so we have "true" Content objects in tests.
Which then requires a few adjustments in journal tests because the
Content's data is never sent to the journal.