client: Add create_topics parameter to JournalClient (for development)
This new parameter enables to create kafka topics if they do not exist when
initializing a journal client.

It should not be set to True in production, its purpose is to delegate topics
creation to journal clients in SWH docker environment instead of creating all
topics prior starting the kafka consumer services. Proceeding like this brings
a huge speedup to initialize the docker compose environment.