Revert "Fix on_eof behavior when the consumer group is rebalanced"
This reverts commit e625c64c6c418aa4be74325bc05b5c5808f22f99 as
the consumers are now lost in an infinite loop after a partition
rebalance.

This commit simply revert the change because the on_eof option
is almost not yet used by any consumer in production (only the journal
scrubber). The proper way to implemennt the change will be
investigate later.

Related to swh/devel/swh-journal#4659